[Backtracing] macOS fixes.
A handful of fixes to get the build going again on macOS after the Linux
work changed a few things.  Most of this is due to rearranging the way
we import the various bits of OS functionality (especially SPI) to not
rely on SwiftShims.

+++ stdlib/public/Backtracing/SymbolicatedBacktrace.swift
@@ -378,7 +378,7 @@ public struct SymbolicatedBacktrace: CustomStringConvertible {
     let theSymbol = Symbol(imageIndex: imageIndex,
                            imageName: imageName,
                            rawName: rawName,
-                           offset: Int(address - UInt(range.location)),
+                           offset: Int(address - UInt64(range.location)),
                            sourceLocation: location)
     theSymbol.name = name
